LayoutTests/fast/regions/cssom/webkit-named-flow-first-empty-region-index-expected.txt

Test for 90608: [CSSRegions]Implement NamedFlow::firstEmptyRegionIndex attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS namedFlow.firstEmptyRegionIndex is -1
PASS namedFlow.firstEmptyRegionIndex is -1
PASS namedFlow.firstEmptyRegionIndex is 1
PASS namedFlow.firstEmptyRegionIndex is 2
PASS namedFlow.firstEmptyRegionIndex is -1
PASS namedFlow.firstEmptyRegionIndex is 0
PASS namedFlow.firstEmptyRegionIndex is -1
PASS successfullyParsed is true

TEST COMPLETE
